refactor: collapse core realtime adapters
