+++ codex-rs/core/src/lib.rs
@@ -11,8 +11,7 @@ mod apps;
 mod arc_monitor;
 mod client;
 mod client_common;
-mod realtime_context;
-mod realtime_conversation;
+mod realtime;
 pub(crate) mod session;
 pub use session::SteerInputError;
 mod codex_thread;
