+++ codex-rs/core/src/session/handlers.rs
@@ -1,7 +1,7 @@
-use crate::realtime_conversation::handle_audio as handle_realtime_conversation_audio;
-use crate::realtime_conversation::handle_close as handle_realtime_conversation_close;
-use crate::realtime_conversation::handle_start as handle_realtime_conversation_start;
-use crate::realtime_conversation::handle_text as handle_realtime_conversation_text;
+use crate::realtime::handle_audio as handle_realtime_conversation_audio;
+use crate::realtime::handle_close as handle_realtime_conversation_close;
+use crate::realtime::handle_start as handle_realtime_conversation_start;
+use crate::realtime::handle_text as handle_realtime_conversation_text;
 use async_channel::Receiver;
 use codex_otel::set_parent_from_w3c_trace_context;
 use codex_protocol::protocol::Submission;
@@ -17,10 +17,10 @@ use crate::config::Config;
 use crate::config_loader::CloudRequirementsLoader;
 use crate::config_loader::LoaderOverrides;
 use crate::config_loader::load_config_layers_state;
-use crate::realtime_context::REALTIME_TURN_TOKEN_BUDGET;
-use crate::realtime_context::truncate_realtime_text_to_token_budget;
-use crate::realtime_conversation::REALTIME_USER_TEXT_PREFIX;
-use crate::realtime_conversation::prefix_realtime_v2_text;
+use crate::realtime::REALTIME_TURN_TOKEN_BUDGET;
+use crate::realtime::REALTIME_USER_TEXT_PREFIX;
+use crate::realtime::prefix_realtime_v2_text;
+use crate::realtime::truncate_realtime_text_to_token_budget;
 use crate::session::spawn_review_thread;
 use codex_exec_server::LOCAL_FS;
 use codex_features::Feature;
